requirements.txt
Flask==2.2.2
gunicorn==20.1.0
google-cloud-secret-manager==2.12.2
google-cloud-storage==2.9.0
fsspec==2023.5.0
gcsfs==2023.5.0
python-gnupg==0.4.7
Werkzeug==2.2.2